backend/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
facenet-pytorch
pillow
numpy
python-multipart
jinja2